feat: visitor queue
